+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/RespondToWrittenRepresentationHandler.java
@@ -15,6 +15,7 @@ import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.common.Element;
 import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.civil.model.documents.Document;
+import uk.gov.hmcts.reform.civil.service.DocUploadDashboardNotificationService;
 import uk.gov.hmcts.reform.civil.service.docmosis.RespondToWrittenRepresentationGenerator;
 import uk.gov.hmcts.reform.civil.utils.DocUploadUtils;
 import uk.gov.hmcts.reform.civil.utils.ElementUtils;
@@ -40,7 +41,7 @@ public class RespondToWrittenRepresentationHandler extends CallbackHandler {
     private final CaseDetailsConverter caseDetailsConverter;
     private final IdamClient idamClient;
     private final RespondToWrittenRepresentationGenerator respondToWrittenRepresentation;
-
+    private final DocUploadDashboardNotificationService docUploadDashboardNotificationService;
     private static final List<CaseEvent> EVENTS = Collections.singletonList(RESPOND_TO_JUDGE_WRITTEN_REPRESENTATION);
 
     @Override
@@ -54,7 +55,8 @@ public class RespondToWrittenRepresentationHandler extends CallbackHandler {
     private CallbackResponse submitClaim(CallbackParams callbackParams) {
 
         CaseData caseData = caseDetailsConverter.toCaseData(callbackParams.getRequest().getCaseDetails());
-        String userId = idamClient.getUserInfo(callbackParams.getParams().get(BEARER_TOKEN).toString()).getUid();
+        String authToken = callbackParams.getParams().get(BEARER_TOKEN).toString();
+        String userId = idamClient.getUserInfo(authToken).getUid();
         CaseData.CaseDataBuilder caseDataBuilder = caseData.toBuilder();
         String role = DocUploadUtils.getUserRole(caseData, userId);
         List<Element<Document>> responseDocumentToBeAdded = caseData.getGeneralAppWrittenRepUpload();
@@ -72,6 +74,9 @@ public class RespondToWrittenRepresentationHandler extends CallbackHandler {
         caseDataBuilder.generalAppWrittenRepText(null);
         caseDataBuilder.businessProcess(BusinessProcess.ready(RESPOND_TO_JUDGE_WRITTEN_REPRESENTATION)).build();
         CaseData updatedCaseData = caseDataBuilder.build();
+
+        // Generate Dashboard Notification for Lip Party
+        docUploadDashboardNotificationService.createDashboardNotification(caseData, role, authToken);
 
         return AboutToStartOrSubmitCallbackResponse.builder()
             .data(updatedCaseData.toMap(objectMapper))
