+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/UploadAdditionalDocumentsCallbackHandler.java
@@ -11,22 +11,15 @@ import uk.gov.hmcts.reform.civil.callback.Callback;
 import uk.gov.hmcts.reform.civil.callback.CallbackHandler;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
-import uk.gov.hmcts.reform.civil.client.DashboardApiClient;
-import uk.gov.hmcts.reform.civil.enums.YesOrNo;
 import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
-import uk.gov.hmcts.reform.civil.launchdarkly.FeatureToggleService;
 import uk.gov.hmcts.reform.civil.model.BusinessProcess;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.common.Element;
 import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.civil.model.genapplication.UploadDocumentByType;
-import uk.gov.hmcts.reform.civil.service.DashboardNotificationsParamsMapper;
-import uk.gov.hmcts.reform.civil.service.GaForLipService;
 import uk.gov.hmcts.reform.civil.utils.AssignCategoryId;
 import uk.gov.hmcts.reform.civil.utils.DocUploadUtils;
 import uk.gov.hmcts.reform.civil.utils.ElementUtils;
-import uk.gov.hmcts.reform.civil.utils.JudicialDecisionNotificationUtil;
-import uk.gov.hmcts.reform.dashboard.data.ScenarioRequestParams;
 import uk.gov.hmcts.reform.idam.client.IdamClient;
 
 import java.time.LocalDateTime;
@@ -40,8 +33,6 @@ import static uk.gov.hmcts.reform.civil.callback.CallbackParams.Params.BEARER_TO
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.SUBMITTED;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.UPLOAD_ADDL_DOCUMENTS;
-import static uk.gov.hmcts.reform.civil.handler.callback.camunda.dashboardnotifications.DashboardScenarios.SCENARIO_OTHER_PARTY_UPLOADED_DOC_APPLICANT;
-import static uk.gov.hmcts.reform.civil.handler.callback.camunda.dashboardnotifications.DashboardScenarios.SCENARIO_OTHER_PARTY_UPLOADED_DOC_RESPONDENT;
 
 @Slf4j
 @Service
@@ -56,10 +47,6 @@ public class UploadAdditionalDocumentsCallbackHandler extends CallbackHandler {
     private final CaseDetailsConverter caseDetailsConverter;
 
     private final IdamClient idamClient;
-    private final DashboardApiClient dashboardApiClient;
-    private final FeatureToggleService featureToggleService;
-    private final GaForLipService gaForLipService;
-    private final DashboardNotificationsParamsMapper mapper;
 
     @Override
     protected Map<String, Callback> callbacks() {
@@ -70,23 +57,16 @@ public class UploadAdditionalDocumentsCallbackHandler extends CallbackHandler {
 
     private CallbackResponse submitDocuments(CallbackParams callbackParams) {
         CaseData caseData = caseDetailsConverter.toCaseData(callbackParams.getRequest().getCaseDetails());
-        String authToken = callbackParams.getParams().get(BEARER_TOKEN).toString();
-        String userId = idamClient.getUserInfo(authToken).getUid();
+        String userId = idamClient.getUserInfo(callbackParams.getParams().get(BEARER_TOKEN).toString()).getUid();
         caseData = buildBundleData(caseData, userId);
         CaseData.CaseDataBuilder caseDataBuilder = caseData.toBuilder();
         String role = DocUploadUtils.getUserRole(caseData, userId);
         DocUploadUtils.addUploadDocumentByTypeToAddl(caseData, caseDataBuilder,
-                                                     caseData.getUploadDocument(), role, true
-        );
+                                                     caseData.getUploadDocument(), role, true);
 
         caseDataBuilder.uploadDocument(null);
         caseDataBuilder.businessProcess(BusinessProcess.ready(UPLOAD_ADDL_DOCUMENTS)).build();
         CaseData updatedCaseData = caseDataBuilder.build();
-
-        if (isWithNoticeOrConsent(caseData)) {
-            createDashboardNotification(caseData, role, authToken);
-        }
-
         return AboutToStartOrSubmitCallbackResponse.builder()
             .data(updatedCaseData.toMap(objectMapper))
             .build();
@@ -111,8 +91,7 @@ public class UploadAdditionalDocumentsCallbackHandler extends CallbackHandler {
             assignCategoryId.assignCategoryIdToCollection(
                 bundle,
                 document -> document.getValue().getDocumentLink(),
-                AssignCategoryId.APPLICATIONS
-            );
+                AssignCategoryId.APPLICATIONS);
             if (Objects.nonNull(caseData.getGaAddlDocBundle())) {
                 bundle.addAll(caseData.getGaAddlDocBundle());
             }
@@ -134,34 +113,4 @@ public class UploadAdditionalDocumentsCallbackHandler extends CallbackHandler {
         return EVENTS;
     }
 
-    private void createDashboardNotification(CaseData caseData, String role, String authToken) {
-
-        if (featureToggleService.isDashboardServiceEnabled() && gaForLipService.isGaForLip(caseData)) {
-            String scenario = getDashboardScenario(role, caseData);
-            ScenarioRequestParams scenarioParams = ScenarioRequestParams.builder().params(mapper.mapCaseDataToParams(
-                caseData)).build();
-            if (scenario != null) {
-                dashboardApiClient.recordScenario(
-                    caseData.getCcdCaseReference().toString(),
-                    scenario,
-                    authToken,
-                    scenarioParams
-                );
-            }
-        }
-    }
-
-    private String getDashboardScenario(String role, CaseData caseData) {
-        if (DocUploadUtils.APPLICANT.equals(role) && gaForLipService.isLipResp(caseData)) {
-            return SCENARIO_OTHER_PARTY_UPLOADED_DOC_RESPONDENT.getScenario();
-        } else if (DocUploadUtils.RESPONDENT_ONE.equals(role) && gaForLipService.isLipApp(caseData)) {
-            return SCENARIO_OTHER_PARTY_UPLOADED_DOC_APPLICANT.getScenario();
-        }
-        return null;
-    }
-
-    private boolean isWithNoticeOrConsent(CaseData caseData) {
-        return JudicialDecisionNotificationUtil.isWithNotice(caseData)
-            || caseData.getGeneralAppConsentOrder() == YesOrNo.YES;
-    }
 }
