+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/user/RespondToJudgeAddlnInfoHandlerTest.java
@@ -24,6 +24,7 @@ import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.civil.model.documents.Document;
 import uk.gov.hmcts.reform.civil.model.genapplication.GAApplicationType;
 import uk.gov.hmcts.reform.civil.model.genapplication.GASolicitorDetailsGAspec;
+import uk.gov.hmcts.reform.civil.service.DocUploadDashboardNotificationService;
 import uk.gov.hmcts.reform.civil.service.docmosis.requestmoreinformation.RespondForInformationGenerator;
 import uk.gov.hmcts.reform.civil.utils.DocUploadUtils;
 import uk.gov.hmcts.reform.idam.client.IdamClient;
@@ -62,6 +63,8 @@ public class RespondToJudgeAddlnInfoHandlerTest extends BaseCallbackHandlerTest 
     IdamClient idamClient;
     @MockBean
     RespondForInformationGenerator respondForInformationGenerator;
+    @MockBean
+    DocUploadDashboardNotificationService docUploadDashboardNotificationService;
 
     private static final String CAMUNDA_EVENT = "INITIATE_GENERAL_APPLICATION";
     private static final String BUSINESS_PROCESS_INSTANCE_ID = "11111";
