+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/user/RespondWrittenRepresentationHandlerTest.java
@@ -25,6 +25,7 @@ import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.civil.model.documents.Document;
 import uk.gov.hmcts.reform.civil.model.genapplication.GAApplicationType;
 import uk.gov.hmcts.reform.civil.model.genapplication.GASolicitorDetailsGAspec;
+import uk.gov.hmcts.reform.civil.service.DocUploadDashboardNotificationService;
 import uk.gov.hmcts.reform.civil.service.docmosis.RespondToWrittenRepresentationGenerator;
 import uk.gov.hmcts.reform.civil.utils.DocUploadUtils;
 import uk.gov.hmcts.reform.idam.client.IdamClient;
@@ -63,6 +64,8 @@ public class RespondWrittenRepresentationHandlerTest extends BaseCallbackHandler
     @MockBean
     RespondToWrittenRepresentationGenerator respondToWrittenRepresentationGenerator;
 
+    @MockBean
+    DocUploadDashboardNotificationService docUploadDashboardNotificationService;
     private static final String CAMUNDA_EVENT = "INITIATE_GENERAL_APPLICATION";
     private static final String BUSINESS_PROCESS_INSTANCE_ID = "11111";
     private static final String ACTIVITY_ID = "anyActivity";
