+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/user/ResponseToJudgeDirectionsOrderHanderTest.java
@@ -25,6 +25,7 @@ import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.civil.model.documents.Document;
 import uk.gov.hmcts.reform.civil.model.genapplication.GAApplicationType;
 import uk.gov.hmcts.reform.civil.model.genapplication.GASolicitorDetailsGAspec;
+import uk.gov.hmcts.reform.civil.service.DocUploadDashboardNotificationService;
 import uk.gov.hmcts.reform.civil.utils.DocUploadUtils;
 import uk.gov.hmcts.reform.idam.client.IdamClient;
 import uk.gov.hmcts.reform.idam.client.models.UserInfo;
@@ -58,6 +59,10 @@ public class ResponseToJudgeDirectionsOrderHanderTest extends BaseCallbackHandle
     CaseDetailsConverter caseDetailsConverter;
     @MockBean
     IdamClient idamClient;
+
+    @MockBean
+    DocUploadDashboardNotificationService docUploadDashboardNotificationService;
+
     private static final String CAMUNDA_EVENT = "INITIATE_GENERAL_APPLICATION";
     private static final String BUSINESS_PROCESS_INSTANCE_ID = "11111";
     private static final String ACTIVITY_ID = "anyActivity";
