+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/user/UploadAdditionalDocumentsCallbackHandlerTest.java
@@ -11,12 +11,10 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.SubmittedCallbackResponse;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
-import uk.gov.hmcts.reform.civil.client.DashboardApiClient;
 import uk.gov.hmcts.reform.civil.enums.BusinessProcessStatus;
 import uk.gov.hmcts.reform.civil.enums.YesOrNo;
 import uk.gov.hmcts.reform.civil.handler.callback.BaseCallbackHandlerTest;
 import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
-import uk.gov.hmcts.reform.civil.launchdarkly.FeatureToggleService;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.common.Element;
 import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
@@ -26,33 +24,25 @@ import uk.gov.hmcts.reform.civil.model.genapplication.GASolicitorDetailsGAspec;
 import uk.gov.hmcts.reform.civil.model.genapplication.GAUrgencyRequirement;
 import uk.gov.hmcts.reform.civil.model.genapplication.UploadDocumentByType;
 import uk.gov.hmcts.reform.civil.sampledata.CaseDataBuilder;
-import uk.gov.hmcts.reform.civil.service.DashboardNotificationsParamsMapper;
-import uk.gov.hmcts.reform.civil.service.GaForLipService;
 import uk.gov.hmcts.reform.civil.utils.AssignCategoryId;
-import uk.gov.hmcts.reform.dashboard.data.ScenarioRequestParams;
 import uk.gov.hmcts.reform.idam.client.IdamClient;
 import uk.gov.hmcts.reform.idam.client.models.UserInfo;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 
 import static java.lang.String.format;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.SUBMITTED;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.UPLOAD_ADDL_DOCUMENTS;
 import static uk.gov.hmcts.reform.civil.enums.YesOrNo.NO;
 import static uk.gov.hmcts.reform.civil.enums.YesOrNo.YES;
-import static uk.gov.hmcts.reform.civil.handler.callback.camunda.dashboardnotifications.DashboardScenarios.SCENARIO_OTHER_PARTY_UPLOADED_DOC_APPLICANT;
-import static uk.gov.hmcts.reform.civil.handler.callback.camunda.dashboardnotifications.DashboardScenarios.SCENARIO_OTHER_PARTY_UPLOADED_DOC_RESPONDENT;
 import static uk.gov.hmcts.reform.civil.sampledata.CaseDataBuilder.STRING_CONSTANT;
 import static uk.gov.hmcts.reform.civil.utils.ElementUtils.element;
 
@@ -75,15 +65,6 @@ class UploadAdditionalDocumentsCallbackHandlerTest extends BaseCallbackHandlerTe
     @MockBean
     AssignCategoryId assignCategoryId;
 
-    @MockBean
-    DashboardApiClient dashboardApiClient;
-    @MockBean
-    FeatureToggleService featureToggleService;
-    @MockBean
-    GaForLipService gaForLipService;
-    @MockBean
-    DashboardNotificationsParamsMapper mapper;
-
     List<Element<CaseDocument>> documents = new ArrayList<>();
 
     private static final String DUMMY_EMAIL = "test@gmail.com";
@@ -94,7 +75,6 @@ class UploadAdditionalDocumentsCallbackHandlerTest extends BaseCallbackHandlerTe
         when(idamClient.getUserInfo(anyString())).thenReturn(UserInfo.builder()
                                                                  .sub(DUMMY_EMAIL)
                                                                  .uid(STRING_CONSTANT)
-                                                                 .roles(List.of())
                                                                  .build());
 
     }
@@ -221,9 +201,9 @@ class UploadAdditionalDocumentsCallbackHandlerTest extends BaseCallbackHandlerTe
                                                                               .build()).build()));
             List<Element<GASolicitorDetailsGAspec>> gaApplAddlSolicitors = new ArrayList<>();
             gaApplAddlSolicitors.add(element(GASolicitorDetailsGAspec.builder()
-                                             .id("id1")
-                                             .email(DUMMY_EMAIL)
-                                             .organisationIdentifier("1").build()));
+                                                 .id("id1")
+                                                 .email(DUMMY_EMAIL)
+                                                 .organisationIdentifier("1").build()));
             CaseData caseData = CaseDataBuilder.builder()
                 .atStateClaimDraft()
                 .ccdCaseReference(1678356749555475L)
@@ -292,11 +272,11 @@ class UploadAdditionalDocumentsCallbackHandlerTest extends BaseCallbackHandlerTe
             List<Element<UploadDocumentByType>> uploadDocumentByRespondent = new ArrayList<>();
             uploadDocumentByRespondent.add(element(UploadDocumentByType.builder()
                                                        .documentType("witness")
-                                                      .additionalDocument(Document.builder()
-                                                                              .documentFileName("witness_document.pdf")
-                                                                              .documentUrl("http://dm-store:8080")
-                                                                              .documentBinaryUrl("http://dm-store:8080/documents")
-                                                                              .build()).build()));
+                                                       .additionalDocument(Document.builder()
+                                                                               .documentFileName("witness_document.pdf")
+                                                                               .documentUrl("http://dm-store:8080")
+                                                                               .documentBinaryUrl("http://dm-store:8080/documents")
+                                                                               .build()).build()));
             List<Element<GASolicitorDetailsGAspec>> gaRespSolicitors = new ArrayList<>();
             gaRespSolicitors.add(element(GASolicitorDetailsGAspec.builder()
                                              .id(STRING_CONSTANT)
@@ -534,37 +514,37 @@ class UploadAdditionalDocumentsCallbackHandlerTest extends BaseCallbackHandlerTe
 
             List<Element<UploadDocumentByType>> uploadDocumentByApplicant = new ArrayList<>();
             uploadDocumentByApplicant.add(element(UploadDocumentByType.builder()
-                    .documentType("bundle")
-                    .additionalDocument(Document.builder()
-                            .documentFileName("witness_document.pdf")
-                            .documentUrl("http://dm-store:8080")
-                            .documentBinaryUrl("http://dm-store:8080/documents")
-                            .build()).build()));
+                                                      .documentType("bundle")
+                                                      .additionalDocument(Document.builder()
+                                                                              .documentFileName("witness_document.pdf")
+                                                                              .documentUrl("http://dm-store:8080")
+                                                                              .documentBinaryUrl("http://dm-store:8080/documents")
+                                                                              .build()).build()));
 
             List<Element<GASolicitorDetailsGAspec>> gaRespSolicitors = new ArrayList<>();
             gaRespSolicitors.add(element(GASolicitorDetailsGAspec.builder()
-                    .id("222")
-                    .email(DUMMY_EMAIL)
-                    .organisationIdentifier("2").build()));
+                                             .id("222")
+                                             .email(DUMMY_EMAIL)
+                                             .organisationIdentifier("2").build()));
             gaRespSolicitors.add(element(GASolicitorDetailsGAspec.builder()
-                    .id(STRING_CONSTANT)
-                    .email(DUMMY_EMAIL)
-                    .organisationIdentifier("3").build()));
+                                             .id(STRING_CONSTANT)
+                                             .email(DUMMY_EMAIL)
+                                             .organisationIdentifier("3").build()));
             CaseData caseData = CaseDataBuilder.builder()
-                    .atStateClaimDraft()
-                    .ccdCaseReference(1678356749555475L)
-                    .build().toBuilder()
-                    .respondent2SameLegalRepresentative(YesOrNo.NO)
-                    .generalAppConsentOrder(YES)
-                    .generalAppRespondentSolicitors(gaRespSolicitors)
-                    .parentClaimantIsApplicant(NO)
-                    .isMultiParty(YES)
-                    .generalAppApplnSolicitor(GASolicitorDetailsGAspec.builder().id("2").forename("GAApplnSolicitor")
-                            .email(DUMMY_EMAIL).organisationIdentifier("1").build())
-                    .uploadDocument(uploadDocumentByApplicant)
-                    .claimant1PartyName("Mr. John Rambo")
-                    .defendant1PartyName("Mr. Sole Trader")
-                    .build();
+                .atStateClaimDraft()
+                .ccdCaseReference(1678356749555475L)
+                .build().toBuilder()
+                .respondent2SameLegalRepresentative(YesOrNo.NO)
+                .generalAppConsentOrder(YES)
+                .generalAppRespondentSolicitors(gaRespSolicitors)
+                .parentClaimantIsApplicant(NO)
+                .isMultiParty(YES)
+                .generalAppApplnSolicitor(GASolicitorDetailsGAspec.builder().id("2").forename("GAApplnSolicitor")
+                                              .email(DUMMY_EMAIL).organisationIdentifier("1").build())
+                .uploadDocument(uploadDocumentByApplicant)
+                .claimant1PartyName("Mr. John Rambo")
+                .defendant1PartyName("Mr. Sole Trader")
+                .build();
             when(caseDetailsConverter.toCaseData(any())).thenReturn(caseData);
             CallbackParams params = callbackParamsOf(caseData, ABOUT_TO_SUBMIT);
             var response = (AboutToStartOrSubmitCallbackResponse) handler.handle(params);
@@ -574,139 +554,6 @@ class UploadAdditionalDocumentsCallbackHandlerTest extends BaseCallbackHandlerTe
             assertThat(responseCaseData.getIsDocumentVisible()).isEqualTo(YesOrNo.YES);
             assertThat(responseCaseData.getGaAddlDocRespondentSolTwo()).isNull();
             assertThat(responseCaseData.getGaAddlDocBundle().size()).isEqualTo(1);
-        }
-
-        @Test
-        void shouldCreateDashboardNotificationWhenLipApplicantUploadDoc() {
-
-            List<Element<UploadDocumentByType>> uploadDocumentByApplicant = new ArrayList<>();
-            uploadDocumentByApplicant.add(element(UploadDocumentByType.builder()
-                                                      .documentType("Witness")
-                                                      .additionalDocument(Document.builder()
-                                                                              .documentFileName("witness_document.pdf")
-                                                                              .documentUrl("http://dm-store:8080")
-                                                                              .documentBinaryUrl("http://dm-store:8080/documents")
-                                                                              .build()).build()));
-            CaseData caseData = CaseDataBuilder.builder()
-                .atStateClaimDraft()
-                .ccdCaseReference(1678356749555475L)
-                .build().toBuilder()
-                .respondent2SameLegalRepresentative(YesOrNo.NO)
-                .applicationIsUncloakedOnce(YES)
-                .parentClaimantIsApplicant(YES)
-                .uploadDocument(uploadDocumentByApplicant)
-                .claimant1PartyName("Mr. John Rambo")
-                .defendant1PartyName("Mr. Sole Trader")
-                .generalAppConsentOrder(YES)
-                .isGaApplicantLip(YES)
-                .isGaRespondentOneLip(YES)
-                .generalAppApplnSolicitor(GASolicitorDetailsGAspec.builder().id(STRING_CONSTANT).forename("GAApplnSolicitor")
-                                              .email(DUMMY_EMAIL).organisationIdentifier("1").build())
-
-                .build();
-
-            HashMap<String, Object> scenarioParams = new HashMap<>();
-            when(gaForLipService.isGaForLip(any(CaseData.class))).thenReturn(true);
-            when(gaForLipService.isLipResp(any(CaseData.class))).thenReturn(true);
-            when(featureToggleService.isDashboardServiceEnabled()).thenReturn(true);
-            when(mapper.mapCaseDataToParams(any())).thenReturn(scenarioParams);
-            when(caseDetailsConverter.toCaseData(any())).thenReturn(caseData);
-            CallbackParams params = callbackParamsOf(caseData, ABOUT_TO_SUBMIT);
-            handler.handle(params);
-            verify(dashboardApiClient).recordScenario(
-                caseData.getCcdCaseReference().toString(),
-                SCENARIO_OTHER_PARTY_UPLOADED_DOC_RESPONDENT.getScenario(),
-                "BEARER_TOKEN",
-                ScenarioRequestParams.builder().params(scenarioParams).build()
-            );
-        }
-
-        @Test
-        void shouldCreateDashboardNotificationWhenLipRespondentUploadDoc() {
-
-            List<Element<UploadDocumentByType>> uploadDocumentByApplicant = new ArrayList<>();
-            uploadDocumentByApplicant.add(element(UploadDocumentByType.builder()
-                                                      .documentType("Witness")
-                                                      .additionalDocument(Document.builder()
-                                                                              .documentFileName("witness_document.pdf")
-                                                                              .documentUrl("http://dm-store:8080")
-                                                                              .documentBinaryUrl("http://dm-store:8080/documents")
-                                                                              .build()).build()));
-            List<Element<GASolicitorDetailsGAspec>> gaRespSolicitors = new ArrayList<>();
-            gaRespSolicitors.add(element(GASolicitorDetailsGAspec.builder()
-                                             .id(STRING_CONSTANT)
-                                             .email(DUMMY_EMAIL)
-                                             .organisationIdentifier("2").build()));
-
-            CaseData caseData = CaseDataBuilder.builder()
-                .atStateClaimDraft()
-                .ccdCaseReference(1678356749555475L)
-                .build().toBuilder()
-                .respondent2SameLegalRepresentative(YesOrNo.NO)
-                .applicationIsUncloakedOnce(YES)
-                .parentClaimantIsApplicant(YES)
-                .uploadDocument(uploadDocumentByApplicant)
-                .claimant1PartyName("Mr. John Rambo")
-                .defendant1PartyName("Mr. Sole Trader")
-                .generalAppConsentOrder(YES)
-                .isGaApplicantLip(YES)
-                .isGaRespondentOneLip(YES)
-                .generalAppApplnSolicitor(GASolicitorDetailsGAspec.builder().id("123456789").forename("GAApplnSolicitor")
-                                              .email(DUMMY_EMAIL).organisationIdentifier("1").build())
-                .generalAppRespondentSolicitors(gaRespSolicitors)
-
-                .build();
-
-            HashMap<String, Object> scenarioParams = new HashMap<>();
-            when(gaForLipService.isGaForLip(any(CaseData.class))).thenReturn(true);
-            when(gaForLipService.isLipApp(any(CaseData.class))).thenReturn(true);
-            when(featureToggleService.isDashboardServiceEnabled()).thenReturn(true);
-            when(mapper.mapCaseDataToParams(any())).thenReturn(scenarioParams);
-            when(caseDetailsConverter.toCaseData(any())).thenReturn(caseData);
-            CallbackParams params = callbackParamsOf(caseData, ABOUT_TO_SUBMIT);
-            handler.handle(params);
-            verify(dashboardApiClient).recordScenario(
-                caseData.getCcdCaseReference().toString(),
-                SCENARIO_OTHER_PARTY_UPLOADED_DOC_APPLICANT.getScenario(),
-                "BEARER_TOKEN",
-                ScenarioRequestParams.builder().params(scenarioParams).build()
-            );
-        }
-
-        @Test
-        void shouldNotCreateDashboardNotificationWhenGaFlagIsDisabled() {
-
-            List<Element<UploadDocumentByType>> uploadDocumentByApplicant = new ArrayList<>();
-            uploadDocumentByApplicant.add(element(UploadDocumentByType.builder()
-                                                      .documentType("Witness")
-                                                      .additionalDocument(Document.builder()
-                                                                              .documentFileName("witness_document.pdf")
-                                                                              .documentUrl("http://dm-store:8080")
-                                                                              .documentBinaryUrl("http://dm-store:8080/documents")
-                                                                              .build()).build()));
-            CaseData caseData = CaseDataBuilder.builder()
-                .atStateClaimDraft()
-                .ccdCaseReference(1678356749555475L)
-                .build().toBuilder()
-                .respondent2SameLegalRepresentative(YesOrNo.NO)
-                .applicationIsUncloakedOnce(YES)
-                .parentClaimantIsApplicant(YES)
-                .uploadDocument(uploadDocumentByApplicant)
-                .claimant1PartyName("Mr. John Rambo")
-                .defendant1PartyName("Mr. Sole Trader")
-                .generalAppConsentOrder(YES)
-                .isGaApplicantLip(YES)
-                .isGaRespondentOneLip(YES)
-                .generalAppApplnSolicitor(GASolicitorDetailsGAspec.builder().id(STRING_CONSTANT).forename("GAApplnSolicitor")
-                                              .email(DUMMY_EMAIL).organisationIdentifier("1").build())
-
-                .build();
-            when(gaForLipService.isGaForLip(any(CaseData.class))).thenReturn(false);
-            when(featureToggleService.isDashboardServiceEnabled()).thenReturn(true);
-            when(caseDetailsConverter.toCaseData(any())).thenReturn(caseData);
-            CallbackParams params = callbackParamsOf(caseData, ABOUT_TO_SUBMIT);
-            handler.handle(params);
-            verifyNoInteractions(dashboardApiClient);
         }
     }
